requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
httpx>=0.24.0
websockets>=11.0.0
typing-extensions>=4.8.0
pandas>=2.0.0
plotly>=5.17.0 